Add verbosity control and update diagnostic logging
Summary:
* Added environment variable `TTTRLIB_VERBOSE` to enable/disable verbosity with flexible options.
* Updated verbose logging across multiple components to use `is_verbose()` for cleaner code.
* Replaced preprocessor-based verbosity checks (`#ifdef VERBOSE_TTTRLIB`) with runtime checks.
* Included new documentation (`verbosity.rst`) to explain toggling verbosity.

+++ include/TTTR.h
@@ -941,7 +941,7 @@ public:
         }
         ranges_by_time_window(
                 output, n_output,
-                macro_times, n_valid_events,
+                macro_times, static_cast<int>(n_valid_events),
                 minimum_window_length,
                 maximum_window_length,
                 minimum_number_of_photons_in_time_window,
